GameOptionsMenuPatchクラスをGameSettingMenuPatchクラスと分離
